APIs call function refactored.
